chore(aws): remove old AWS provider
